backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
sqlalchemy==2.0.21
passlib==1.7.4
pyjwt==2.8.0
email-validator==2.0.0
pytest==7.4.2
Werkzeug==2.3.7
